# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Practice API Documentation",
        "description": "Practice API description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of Users",
                "description": "Get list of Users",
                "operationId": "fetchUsers",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "search between name and email of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "PaginatedUser",
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "title": "User",
                                                    "type": "array",
                                                    "items": {
                                                        "allOf": [
                                                            {
                                                                "$ref": "#/components/schemas/User"
                                                            },
                                                            {
                                                                "$ref": "#/components/schemas/CreatedAndUpdatedDates"
                                                            }
                                                        ]
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "create a new Users",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "CreatedUser",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "$ref": "#/components/schemas/CreatedAndUpdatedDates"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/users/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "update a Users",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "UpdatedUser",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "$ref": "#/components/schemas/CreatedAndUpdatedDates"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "delete Users",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "number"
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "number"
                    },
                    "last_page": {
                        "type": "number"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "links": {
                        "title": "PaginationLinks",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "number"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "number"
                    },
                    "total": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "CreatedAndUpdatedDates": {
                "properties": {
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "example@example.com"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the user",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the user",
                        "format": "email",
                        "example": "example@example.com"
                    },
                    "email_verified_at": {
                        "description": "Email verification",
                        "format": "date",
                        "example": "2023-01-01 09:32:29"
                    },
                    "password": {
                        "description": "Password of the user",
                        "example": "password"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Practice",
            "description": "API Endpoints of Practice"
        }
    ]
}